+++ src/utils/botUtilsConstants.js
@@ -15,8 +15,8 @@ const HELP_MESSAGE = 'This is what I can do: \n'
     + '* `run <command>`: I\'ll run a command on the server\n'
     + '* `wallet <username>`: I\'ll check how much money you have in your wallet\n'
     + '* `price <itemName>`: I\'ll price check an item from the Shop\n'
-    + '* `buy <username> <quantity> <itemName>`: I\'ll buy you items from the Shop\n'
-    + '* `sell <username> <quantity> <itemName>`: I\'ll sell your items to the Shop\n'
+    + '* `buy <quantity> <itemName>`: I\'ll buy you items from the Shop\n'
+    + '* `sell <quantity> <itemName>`: I\'ll sell your items to the Shop\n'
     + '* `help`: Show this help, since as it stands you can\'t do anything for yourself';
 
 const EMPTY_SEARCH_RESULTS_MESSAGE = "No results!";